% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_eml.R
\name{as_eml}
\alias{as_eml}
\title{Crosswalk \code{dataspice} metadata to EML}
\usage{
as_eml(spice)
}
\arguments{
\item{spice}{(list) Your \code{dataspice} metadata}
}
\value{
(emld) The crosswalked \code{emld} object
}
\description{
Performs an (imperfect) crosswalk of \code{dataspice} metadata to EML 2.2.0. It's
very likely you will get validation errors and need to fix them afterwards
but \code{as_emld} is a good way to a richer metadata schema (EML) when you're
already using \code{dataspice}.
}
\examples{
# Load an example dataspice JSON that comes installed with the package
spice <- system.file(
  "examples", "annual-escapement.json",
  package = "dataspice")

# And crosswalk it to EML
as_eml(spice)

# We can also create dataspice metadata from scratch and crosswalk it to EML
myspice <- list(
  name = "My example spice",
  creator = "Me",
  contact = "Me")
as_eml(myspice)
}
