% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_emld.R
\name{as_emld}
\alias{as_emld}
\title{Convert \code{dataspice} metadata to EML}
\usage{
as_emld(spice, verbose = FALSE)
}
\arguments{
\item{spice}{(list)}

\item{verbose}{(logical) Whether or not to display extra information during
conversion. Defaults to \code{FALSE}.}
}
\value{
(emld) The conerted \code{emld} object
}
\description{
Performs an imperfect conversion of \code{dataspice} metadata to EML 2.2.0. It's
very likely you will get validation errors and need to fix them afterwards.
}
\examples{
spice <- system.file("examples", "annual-escapement.json", package = "dataspice")
as_emld(spice)
}
