% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jsonld.R
\name{as_jsonld}
\alias{as_jsonld}
\title{Convert a list object to JSON-LD}
\usage{
as_jsonld(x, context = "http://schema.org", pretty = TRUE,
  auto_unbox = TRUE, ...)
}
\arguments{
\item{x}{the object to be encoded}

\item{context}{JSON-LD context; "http://schema.org"}

\item{pretty}{adds indentation whitespace to JSON output. Can be TRUE/FALSE or a number specifying the number of spaces to indent. See \code{\link{prettify}}}

\item{auto_unbox}{automatically \code{\link{unbox}} all atomic vectors of length 1. It is usually safer to avoid this and instead use the \code{\link{unbox}} function to unbox individual elements.
An exception is that objects of class \code{AsIs} (i.e. wrapped in \code{I()}) are not automatically unboxed. This is a way to mark single values as length-1 arrays.}

\item{...}{arguments passed on to class specific \code{print} methods}
}
\description{
Convert a list object to JSON-LD
}
\examples{

x <- Thing(url = "http://schema.org")
as_jsonld(x)

}
