% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spice.R
\name{create_spice}
\alias{create_spice}
\title{Put metadata templates within a \code{metadata} subdirectory}
\usage{
create_spice(dir)
}
\arguments{
\item{dir}{Directory containing data, within which a \code{metadata} subdirectory
will be created}
}
\description{
Put metadata templates within a \code{metadata} subdirectory
}
\examples{
\dontrun{
create_spice(dir = "data")
}
}
