% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_crosswalk.R
\name{crosswalk}
\alias{crosswalk}
\title{Crosswalk a term}
\usage{
crosswalk(doc, term, verbose = FALSE)
}
\arguments{
\item{doc}{(list) A \code{dataspice} document as a \code{list}}

\item{term}{(character) The term to crosswalk}

\item{verbose}{(logical) Whether or not to display extra information during
conversion. Defaults to \code{FALSE}.}
}
\value{
(list) The result of the crosswalk. May be an empty \code{list} on
failure.
}
\description{
Crosswalk a term
}
