% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_crosswalk.R
\name{crosswalk_creator}
\alias{crosswalk_creator}
\title{Crosswalk a Schema.org/creator}
\usage{
crosswalk_creator(creator)
}
\arguments{
\item{creator}{(list) A creator}
}
\description{
Crosswalk a Schema.org/creator
}
