% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_crosswalk.R
\name{crosswalk_datetime}
\alias{crosswalk_datetime}
\title{Convert a date(time) of unknown format into EML}
\usage{
crosswalk_datetime(input)
}
\arguments{
\item{input}{(character) Some unknown date(time) input}
}
\value{
(list) A \code{list} with members \code{calendarDate} and \code{time}. \code{time} will
be \code{NULL} if parsing fails or if the time string inside \code{input} isn't
ISO8601
}
\description{
A quick and dirty crosswalk of an unknown date(time) input to EML that really
only works for ISO8601 input. All other formats will fail and be returned
as-is as a \code{calendarDate}. From there the user will need to do a conversion
themselves.
}
