% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_crosswalk.R
\name{crosswalk_distribution}
\alias{crosswalk_distribution}
\title{Crosswalk a Schema.org/distribution}
\usage{
crosswalk_distribution(distribution)
}
\arguments{
\item{distribution}{(list) A distribution}
}
\description{
Crosswalk a Schema.org/distribution
}
