% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_crosswalk.R
\name{crosswalk_Organization}
\alias{crosswalk_Organization}
\title{Crosswalk a Schema.org/Organization}
\usage{
crosswalk_Organization(creator)
}
\arguments{
\item{creator}{(list) A creator}
}
\description{
Crosswalk a Schema.org/Organization
}
