% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_crosswalk.R
\name{crosswalk_Person}
\alias{crosswalk_Person}
\title{Crosswalk functions for \code{as_eml}
Crosswalk a Schema.org/Person}
\usage{
crosswalk_Person(creator)
}
\arguments{
\item{creator}{(list) A creator}
}
\description{
Crosswalk functions for \code{as_eml}
Crosswalk a Schema.org/Person
}
