% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_access.R
\name{edit_access}
\alias{edit_access}
\title{Shiny App for editing the metadata access table}
\usage{
edit_access(DF, outdir = getwd(), outfilename = "access")
}
\arguments{
\item{DF}{the imported access.csv dataframe}

\item{outdir}{The directory to save the edited access info to}

\item{outfilename}{The filename to save with. Defaults to access.csv.}
}
\description{
Shiny App for editing the metadata access table
}
\examples{
\dontrun{
editTable(DF = access)

}
}
