% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_attributes.R
\name{edit_attributes}
\alias{edit_attributes}
\title{Shiny App for editing the metadata attributes table}
\usage{
edit_attributes(filepath = here::here("data", "metadata", "attributes.csv"))
}
\arguments{
\item{filepath}{the filepath to the dataspice attributes.csv file. Defaults to current
<project_root>/data/metadata/attributes.csv,}
}
\description{
Shiny App for editing the metadata attributes table
}
\examples{
\dontrun{
edit_attributes()

}
}
