% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_biblio.r
\name{edit_biblio}
\alias{edit_biblio}
\title{Shiny App for editing the metadata biblio table}
\usage{
edit_biblio(filepath = here::here("data", "metadata", "biblio.csv"))
}
\arguments{
\item{filepath}{the filepath to the dataspice biblio.csv file. Defaults to current
<project_root>/data/metadata/biblio.csv.}
}
\description{
Shiny App for editing the metadata biblio table
}
\examples{
\dontrun{
edit_biblio()

}
}
