% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_creators.R
\name{edit_creators}
\alias{edit_creators}
\title{Shiny App for editing the metadata creator table}
\usage{
edit_creators(filepath = "metadata-tables/creators.csv", outdir = getwd(),
  outfilename = "creator", numCreators = 10)
}
\arguments{
\item{filepath}{the filepath leading to the creators.csv file}

\item{outdir}{The directory to save the edited creator info to}

\item{outfilename}{The filename to save with. Defaults to creator.csv.}

\item{numCreators}{the number of creators that need to be included, defauls to 10}
}
\description{
Shiny App for editing the metadata creator table
}
\examples{
\dontrun{
editTable(DF = creator)

}
}
