% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editBiblio.r
\name{editBiblio}
\alias{editBiblio}
\title{Shiny App for editing the metadata biblio table}
\usage{
editBiblio(DF, outdir = getwd(), outfilename = "biblio")
}
\arguments{
\item{DF}{the imported biblio.csv dataframe}
}
\description{
Shiny App for editing the metadata biblio table
}
\examples{
\dontrun{
editTable(DF = attributes)

}
}
