% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editCreators.R
\name{editCreators}
\alias{editCreators}
\title{Shiny App for editing the metadata creator table}
\usage{
editCreators(DF, outdir = getwd(), outfilename = "creator",
  numCreators = 10)
}
\arguments{
\item{DF}{the imported creator.csv dataframe}

\item{numCreators}{the number of creators that need to be included, defauls to 10}
}
\description{
Shiny App for editing the metadata creator table
}
\examples{
\dontrun{
editTable(DF = creator)

}
}
