% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_spice.R
\name{es_access}
\alias{es_access}
\title{Get access from EML}
\usage{
es_access(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) folder path for saving the table to disk}
}
\description{
Return EML access in the dataspice access.csv format.
}
