% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_spice.R
\name{es_attributes}
\alias{es_attributes}
\title{Get attributes from EML}
\usage{
es_attributes(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) folder path for saving the table to disk}
}
\description{
Return EML attributes in the dataspice attributes.csv format.
}
