% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_spice.R
\name{es_biblio}
\alias{es_biblio}
\title{Get biblio from EML}
\usage{
es_biblio(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) folder path for saving the table to disk}
}
\description{
Return EML biblio in the dataspice biblio.csv format.
}
