% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_spice.R
\name{es_creators}
\alias{es_creators}
\title{Get creators from EML}
\usage{
es_creators(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) folder path for saving the table to disk}
}
\description{
Return EML creators in the dataspice creators.csv format.
}
