% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonld_to_mustache.R
\name{parse_GeoShape_box}
\alias{parse_GeoShape_box}
\title{Parse spatialCoverage$geo$box section for use in a Leaflet map}
\usage{
parse_GeoShape_box(box)
}
\arguments{
\item{box}{(list) spatialCoverage$geo$box section of the JSONLD}
}
\value{
(list) Template-specific variables for Leaflet
}
\description{
Parse spatialCoverage$geo$box section for use in a Leaflet map
}
