% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonld_to_mustache.R
\name{parse_spatialCoverage}
\alias{parse_spatialCoverage}
\title{Parse spatialCoverge section for use in a Leaflet map}
\usage{
parse_spatialCoverage(spatialCoverage)
}
\arguments{
\item{spatialCoverage}{(list) spatialCoverage section of the JSONLD}
}
\value{
(list) Template-specific variables for Leaflet
}
\description{
Parse spatialCoverge section for use in a Leaflet map
}
