% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_access.R
\name{prep_access}
\alias{prep_access}
\title{Prepare access}
\usage{
prep_access(data_path = here::here("data"),
  access_path = here::here("data", "metadata", "access.csv"))
}
\arguments{
\item{data_path}{path to the data folder. Defaults to "data" and R 'data' file types}

\item{access_path}{path to the access.csv file. Defaults to "data/metadata/access.csv".}
}
\value{
the functions writes out the updated access.csv file to access_path.
}
\description{
Extract variableNames for a given data file and add them to the attributes.csv
}
