% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_attributes}
\alias{prep_attributes}
\title{Prepare attributes}
\usage{
prep_attributes(data_path = here::here("data"),
  attributes_path = here::here("data", "metadata", "attributes.csv"),
  ...)
}
\arguments{
\item{data_path}{character vector of either:
\enumerate{
\item path(s) to the data file(s).
\item single path to directory containing data file(s).
}

Currently only tabular csv and tsv files are supported. Alternatively attributes
returned using \code{names()} can be extracted from an r object, stored as an \code{.rds} file.}

\item{attributes_path}{path to the `attributes.csv`` file. Defaults to "data/metadata/attributes.csv".}

\item{...}{parameters passed to \code{list.files()}. For example, use \code{recursive = TRUE}
to list files in a folder recursively or use \code{pattern} to filter files for patterns.}
}
\value{
the functions writes out the updated attributes.csv file to attributes_path.
}
\description{
Extract `variableNames`` from data file(s) and add them to the attributes.csv
}
