% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve_site.R
\name{serve_site}
\alias{serve_site}
\title{Serve site}
\usage{
serve_site(path = "docs")
}
\arguments{
\item{path}{(character) Optional. Directory to serve. Defaults to \code{docs}.}
}
\value{
Nothing.
}
\description{
Serve site
}
\examples{
\dontrun{
# Build your site
json <- write_json(biblio, access, attributes, creators)
build_site(json)

# Serve it
serve_site()
}
}
