% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_access}
\alias{validate_access}
\title{Validate access.csv}
\usage{
validate_access(access)
}
\arguments{
\item{access}{(data.frame) A \code{data.frame} read in from \code{access.csv}}
}
\value{
Nothing. Side-effect: Can \code{stop} execution if validation fails.
}
\description{
Validate access.csv
}
