% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_biblio}
\alias{validate_biblio}
\title{Validate biblio.csv}
\usage{
validate_biblio(biblio)
}
\arguments{
\item{biblio}{(data.frame) A \code{data.frame} read in from \code{biblio.csv}}
}
\value{
Nothing. Side-effect: Can \code{stop} execution if validation fails.
}
\description{
Validate biblio.csv
}
