% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_creators}
\alias{validate_creators}
\title{Validate creators.csv}
\usage{
validate_creators(creators)
}
\arguments{
\item{creators}{(data.frame) A \code{data.frame} read in from
\code{creators.csv}}
}
\value{
Nothing. Side-effect: Can \code{stop} execution if validation fails.
}
\description{
Validate creators.csv
}
