% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spice.R
\name{write_spice}
\alias{write_spice}
\title{write_spice}
\usage{
write_spice(path = "data/metadata", ...)
}
\arguments{
\item{path}{location of metadata files}

\item{...}{additional arguments to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}}
}
\value{
a json-ld file at the path specified
}
\description{
write_spice
}
