% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_jsonld.R
\name{as_jsonld}
\alias{as_jsonld}
\title{Convert a list object to JSON-LD}
\usage{
as_jsonld(
  x,
  context = "http://schema.org",
  pretty = TRUE,
  auto_unbox = TRUE,
  ...
)
}
\arguments{
\item{x}{the object to be encoded.}

\item{context}{JSON-LD context; "http://schema.org".}

\item{pretty}{Whether or not to prettify output. See
\code{\link[jsonlite]{toJSON}}.}

\item{auto_unbox}{Whether or not to automatically unbox output. See
\code{\link[jsonlite]{toJSON}}.}

\item{...}{Other arguments to be passed to
\code{\link[jsonlite]{toJSON}}.}
}
\description{
Convert a list object to JSON-LD
}
