% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spice.R
\name{create_spice}
\alias{create_spice}
\title{Put metadata templates within a \code{metadata} subdirectory}
\usage{
create_spice(dir = "data")
}
\arguments{
\item{dir}{Directory containing data, within which a \code{metadata} subdirectory
will be created. Defaults to \code{data}.}
}
\description{
Put metadata templates within a \code{metadata} subdirectory
}
\examples{
\dontrun{
create_spice()

# Create templates from the data in a folder other than `data`
create_spice("my_data")
}
}
