% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_access.R
\name{edit_access}
\alias{edit_access}
\title{Shiny App for editing the metadata access table}
\usage{
edit_access(filepath = here::here("data", "metadata", "access.csv"))
}
\arguments{
\item{filepath}{the filepath to the dataspice access.csv file. Defaults to current
<project_root>/data/metadata/access.csv.}
}
\description{
Shiny App for editing the metadata access table
}
\examples{
\dontrun{
edit_access()

}
}
