% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_attributes.R
\name{edit_attributes}
\alias{edit_attributes}
\title{Shiny App for editing the metadata attributes table}
\usage{
edit_attributes(filepath = "metadata-tables/attributes.csv",
  outdir = getwd(), outfilename = "attributes")
}
\arguments{
\item{filepath}{the filepath leading to the attributes.csv file}

\item{outdir}{The directory to save the edited attributes info to}

\item{outfilename}{The filename to save with. Defaults to attributes.csv.}
}
\description{
Shiny App for editing the metadata attributes table
}
\examples{
\dontrun{
editTable(DF = attributes)

}
}
