% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_biblio.r
\name{edit_biblio}
\alias{edit_biblio}
\title{Shiny App for editing the metadata biblio table}
\usage{
edit_biblio(filepath = "metadata-tables/biblio.csv", outdir = getwd(),
  outfilename = "biblio")
}
\arguments{
\item{filepath}{the filepath leading to the biblio.csv file}

\item{outdir}{The directory to save the edited biblio info to}

\item{outfilename}{The filename to save with. Defaults to biblio.csv.}
}
\description{
Shiny App for editing the metadata biblio table
}
\examples{
\dontrun{
editTable(DF = attributes)

}
}
