% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_creators.R
\name{edit_creators}
\alias{edit_creators}
\title{Shiny App for editing the metadata creator table}
\usage{
edit_creators(filepath = here::here("data", "metadata", "creators.csv"),
  numCreators = 10)
}
\arguments{
\item{filepath}{the filepath to the dataspice access.csv file. Defaults to current
<project_root>/data/metadata/access.csv.}

\item{numCreators}{the number of creators that need to be included, defauls to 10}
}
\description{
Shiny App for editing the metadata creator table
}
\examples{
\dontrun{
edit_creators()

}
}
