% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_file.R
\name{edit_creators}
\alias{edit_creators}
\title{Shiny apps for editing dataspice metadata tables}
\usage{
edit_creators(metadata_dir = file.path("data", "metadata"))
}
\arguments{
\item{metadata_dir}{the directory containing the \code{dataspice} metadata \code{.csv}
files. Defaults to \verb{data/metadata/} directory in \strong{current project root}.}
}
\description{
Launch Shiny app for editing individual \code{dataspice} metadata tables. Use
\verb{edit_*()} where \code{*} is one of the four \code{dataspice} metadata tables
\strong{\code{attributes}, \code{biblio}, \code{access} or \code{creators}}.
}
\examples{
\dontrun{
edit_attributes()
edit_biblio()
edit_access()
edit_creators()

# Specifying a different dataspice metadata directory
edit_attributes(metadata_dir = "analysis/data/metadata/"))
}
}
