% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editAccess.R
\name{editAccess}
\alias{editAccess}
\title{Shiny App for editing the metadata access table}
\usage{
editAccess(DF, outdir = getwd(), outfilename = "access")
}
\arguments{
\item{DF}{the imported access.csv dataframe}
}
\description{
Shiny App for editing the metadata access table
}
\examples{
\dontrun{
editTable(DF = access)

}
}
