% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eml_to_spice.R
\name{eml_to_spice}
\alias{eml_to_spice}
\title{Get dataspice tabular formats from EML}
\usage{
eml_to_spice(eml, path = NULL)
}
\arguments{
\item{eml}{(emld) an EML object}

\item{path}{(character) folder path for saving the table to disk}
}
\description{
Return EML in the dataspice dataframes.
}
\examples{
\dontrun{
eml_path <- system.file("example-dataset/broodTable_metadata.xml", package = "dataspice")
eml <- read_eml(eml_path)
my_spice <- eml_to_spice(eml, ".")
}
}
