% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonld_to_mustache.R
\name{jsonld_to_mustache}
\alias{jsonld_to_mustache}
\title{Convert JSONLD to a list suitable for Mustache templating}
\usage{
jsonld_to_mustache(path)
}
\arguments{
\item{path}{(character) Path to file on disk to convert}
}
\value{
(list) Mustache-appropriate list
}
\description{
Convert JSONLD to a list suitable for Mustache templating
}
