% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_attributes}
\alias{prep_attributes}
\title{Prepare attributes}
\usage{
prep_attributes(data_path = here::here("data"),
  attributes_path = here::here("data", "metadata", "attributes.csv"),
  ...)
}
\arguments{
\item{data_path}{path to the data file. Currently only tabular csv and tsv files are supported.}

\item{attributes_path}{path to the attributes.csv file. Defaults to "data/metadata/attributes.csv".}

\item{...}{parameters passed to readr::read_*() functions}
}
\value{
the functions writes out the updated attributes.csv file to attributes_path.
}
\description{
Extract variableNames for a given data file and add them to the attributes.csv
}
