% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_metadata.R
\name{validate_attributes}
\alias{validate_attributes}
\title{Validate attributes.csv}
\usage{
validate_attributes(attributes)
}
\arguments{
\item{attributes}{(data.frame) A \code{data.frame} read in from
\code{attributes.csv}}
}
\value{
Nothing. Side-effect: Can \code{stop} execution if validation fails.
}
\description{
Validate attributes.csv
}
