% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spice.R
\name{write_spice}
\alias{write_spice}
\title{Write spice}
\usage{
write_spice(path = "data/metadata", ...)
}
\arguments{
\item{path}{location of metadata files}

\item{...}{additional arguments to \code{\link[jsonlite:fromJSON]{jsonlite::toJSON()}}}
}
\value{
A JSON-LD file at the path specified
}
\description{
Write out your metadata as a dataspice JSON-LD document
}
\examples{
\dontrun{
# First create your metadata templates
create_spice()

# Then fill in the template files however you like

# Then write out your dataspice file
write_spice()
}
}
